x/alliance/types/messages_quorum_params_test.go
package types

import (
	"testing"

	"kepler/testutil/sample"

	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
	"github.com/stretchr/testify/require"
)

func TestMsgCreateQuorumParams_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgCreateQuorumParams
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgCreateQuorumParams{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgCreateQuorumParams{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgUpdateQuorumParams_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgUpdateQuorumParams
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgUpdateQuorumParams{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgUpdateQuorumParams{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}

func TestMsgDeleteQuorumParams_ValidateBasic(t *testing.T) {
	tests := []struct {
		name string
		msg  MsgDeleteQuorumParams
		err  error
	}{
		{
			name: "invalid address",
			msg: MsgDeleteQuorumParams{
				Creator: "invalid_address",
			},
			err: sdkerrors.ErrInvalidAddress,
		}, {
			name: "valid address",
			msg: MsgDeleteQuorumParams{
				Creator: sample.AccAddress(),
			},
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			err := tt.msg.ValidateBasic()
			if tt.err != nil {
				require.ErrorIs(t, err, tt.err)
				return
			}
			require.NoError(t, err)
		})
	}
}
